refactor PRMSSoilzoneAg to also have PRMSSoilzoneAgObsEt which requires observed aet and pet
